Bocfel: Silence analyzer warnings

+++ terps/bocfel/zterp.c
@@ -1146,7 +1146,7 @@ int main(int argc, char **argv)
             die("unknown story type: %s (0x%08lx)", chunk->name, (unsigned long)chunk->type);
         }
 
-        if (chunk->offset > LONG_MAX) {
+        if (chunk->offset > UINT32_MAX) {
             die("zcode offset too large");
         }
 
@@ -1171,7 +1171,7 @@ int main(int argc, char **argv)
     if (memory_size < 64) {
         die("story file too small");
     }
-    if (memory_size > SIZE_MAX - 22) {
+    if (memory_size > UINT32_MAX - 22) {
         die("story file too large");
     }
